Random Number Generator (#2)
Created a Random Number Generator.

Created a pseudo random number generator that uses a floating analog input pin to generate random numbers. The substituted the RNG into the SnakeGame in place of previous method for generating random numbers.

Additionally, reworked how the ADC is stored, since I had incorrectly assumed that the ADC was a software component that could be freely generated, not a singleton.